src/test/java/com/kp/leetcode/NumOfIslandsTest.java
package com.kp.leetcode;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class NumOfIslandsTest {

    @Test
    void numIslands() {

        char[][] grid = {
                {'1', '1', '1', '1', '0'},
                {'1', '1', '0', '1', '0'},
                {'1', '1', '0', '0', '0'},
                {'0', '0', '0', '0', '0'}
        };

        assertEquals(1,new NumOfIslands().numIslands(grid));
    }


    @Test
    void numIslands2() {

        char[][] grid = {
                {'1', '1', '0', '0', '0'},
                {'1', '1', '0', '0', '0'},
                {'0', '0', '1', '0', '0'},
                {'0', '0', '0', '1', '1'}
        };

        assertEquals(3,new NumOfIslands().numIslands(grid));
    }

    @Test
    void numIslands3() {

        char[][] grid = {
                {'1', '1', '1'},
                {'0', '1', '0'},
                {'1', '1', '1'}
        };

        assertEquals(1,new NumOfIslands().numIslands(grid));
    }

}